chore(allow-scripts): refactor: break apart index.js
This breaks apart the monolothic index.js of allow-scripts into a couple
of internal submodules. Common types that would need to be shared are
broken out as package global types. No functional changes.

index.js is still exposing all functions like before.

This allows more selective import and bundling of allow-scripts
functinonality without needing to consider as many dependencies.

- types/common.ts: Common and exposed types broken out to TS globals
- setup.js: functions related to reading and updating package configuration
- report.js: functions printing results to stdout
- runAllowedPackages.js: The runAllowedPackages function
