+++ packages/allow-scripts/src/cli.js
@@ -2,13 +2,14 @@
 
 const yargs = require('yargs/yargs')
 const { hideBin } = require('yargs/helpers')
+const runAllowedPackages = require('./runAllowedPackages.js')
+const { printPackagesList } = require('./report.js')
 const {
-  runAllowedPackages,
+  editPackageJson,
   setDefaultConfiguration,
-  printPackagesList,
-} = require('./index.js')
-const { writeRcFile, editPackageJson } = require('./setup.js')
-const { FEATURE } = require('./toggles')
+  writeRcFile,
+} = require('./setup.js')
+const { FEATURE } = require('./toggles.js')
 
 start().catch((err) => {
   console.error(err)
